+++ release_notes.md
@@ -6,6 +6,8 @@ short summary of each release, with version and release date.
 work in progress.
 this work is available on "master" branch, but has not yet been released.
 
+  * `Container::ArrayWithSize::back()` added, for easier access.
+  * `Container::ArrayWithSize::front()` added, for easier access.
   * `Container::BitPackedSequence::resize()` added (2 overloads, as in `std::vector`).
 
 ## v1.1.0 (2018-10-10)